rfind returns size_t not unsigned int. Found from static analysis.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/utils@120 720011fe-810a-0410-ac5b-998ee3860a72

+++ modelloader/BVHDataLoader.cpp
@@ -76,7 +76,7 @@ void CBVHDataInstance::load(string file, double radius, double height) {
       if( first == "keyframe" ) { //load keyframe
 	 ILoadable * model=NULL;
 	 //get extension
-	 unsigned int pos=file.rfind('.');
+	 size_t pos=file.rfind('.');
 	 if( pos==string::npos ){
 	    cerr<<"! Error : Can't Recognize file :"<<file<<endl;
 	    exit(0);
